added Solid Layer support

+++ src/Model/Layers/solid_layer.cpp
@@ -23,8 +23,8 @@ SolidLayer::~SolidLayer()
 void SolidLayer::decode(QJsonObject &in_obj, QList<QString> &out_messages)
 {
     m_color = QColor::fromString(in_obj.take(color_key).toString());
-    m_width = in_obj.take(width_key).toDouble();
-    m_height = in_obj.take(height_key).toDouble();
+    m_width = in_obj.take(width_key).toInt();
+    m_height = in_obj.take(height_key).toInt();
     LayerModel::decode(in_obj, out_messages);
 
     for(const auto& element : in_obj.keys()) {
